requirements.txt
"""
Requirements file for Emotion Detection + Response Generation Project
Install with: pip install -r requirements.txt
"""

# Core deep learning
torch>=2.0.0
transformers>=4.30.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Response Generation (Phase 4)
groq>=0.4.0

# Environment Variables
python-dotenv>=1.0.0

# Database (Phase 5)
pymongo>=4.6.0
bcrypt>=4.1.0

# Analytics & Monitoring (Phase 7)
plotly>=5.17.0
psutil>=5.9.0

# Optional: For faster training
# accelerate>=0.20.0
